rename modal file
